+++ src/push/type/item/ref.clj
@@ -3,6 +3,7 @@
             [push.type.core :as t]
             [push.instructions.dsl :as d]
             [push.instructions.aspects :as aspects]
+            [push.util.numerics :as num]
             ))
 
 
@@ -70,6 +71,28 @@
 
 
 
+
+(def ref-cyclevector
+  (core/build-instruction
+    ref-cyclevector
+    "`:ref-cyclevector` pops the top `:ref` item and the top two `:scalar` items (call them `scale` and `raw-count`, respectively). The `scale` value is used to determine whether to convert `raw-count` into a :few, :some, :many or :lots value, and then the appropriate number of elements from the `:ref`'s current stack are made into a single vector (by cycling) and pushed to `:exec`. If the `:ref` has no bound values, an empty vector is pushed."
+    :tags #{:binding}
+    (d/consume-top-of :ref :as :which)
+    (d/save-binding-stack :which :as :contents)
+    (d/consume-top-of :scalar :as :scale)
+    (d/consume-top-of :scalar :as :raw-count)
+    (d/calculate [:scale]
+      #(nth [10 100 1000 10000] (num/scalar-to-index %1 4)) :as :relative)
+    (d/calculate [:raw-count :relative]
+      #(num/scalar-to-index %1 %2) :as :size)
+    (d/calculate [:contents :size]
+      #(into [] (take %2 (cycle %1))) :as :result)
+    (d/push-onto :exec :result)
+    ))
+
+
+
+
 (def ref-dump
   (core/build-instruction
     ref-dump
@@ -93,6 +116,29 @@
     (d/replace-binding :values1 :into :arg2)
     (d/replace-binding :values2 :into :arg1)
     ))
+
+
+
+(def ref-fillvector
+  (core/build-instruction
+    ref-fillvector
+    "`:ref-fillvector` pops the top `:ref` item and the top two `:scalar` items (call them `scale` and `raw-count`, respectively). The `scale` value is used to determine whether to convert `raw-count` into a :few, :some, :many or :lots value, and then the appropriate number of copies of the `:ref`'s current value are made into a single vector and pushed to `:exec`. If the `:ref` has no bound values, an empty vector is pushed."
+    :tags #{:binding}
+    (d/consume-top-of :ref :as :which)
+    (d/save-top-of-binding :which :as :item)
+    (d/consume-top-of :scalar :as :scale)
+    (d/consume-top-of :scalar :as :raw-count)
+    (d/calculate [:scale]
+      #(nth [10 100 1000 10000] (num/scalar-to-index %1 4)) :as :relative)
+    (d/calculate [:raw-count :relative]
+      #(num/scalar-to-index %1 %2) :as :size)
+    (d/calculate [:item :size]
+      #(if (nil? %1)
+        []
+        (into [] (take %2 (repeat %1)))) :as :result)
+    (d/push-onto :exec :result)
+    ))
+
 
 
 
@@ -162,6 +208,18 @@
 
 
 
+(def ref->vector
+  (core/build-instruction
+    ref->vector
+    "`:ref->vector` pops the top `:ref` keyword and copies its entire stack of contents into a new `:vector` item, which is pushed to `:exec`"
+    :tags #{:binding}
+    (d/consume-top-of :ref :as :arg)
+    (d/save-binding-stack :arg :as :contents)
+    (d/calculate [:contents] #(into [] %1) :as :result)
+    (d/push-onto :exec :result)
+    ))
+
+
 
 
 
@@ -178,14 +236,17 @@
 
         (t/attach-instruction ref-ARGS)
         (t/attach-instruction ref-clear)
+        (t/attach-instruction ref-cyclevector)
         (t/attach-instruction ref-dump)
         (t/attach-instruction ref-exchange)
+        (t/attach-instruction ref-fillvector)
         (t/attach-instruction ref-forget)
         (t/attach-instruction ref-fullquote)
         (t/attach-instruction ref-known?)
         (t/attach-instruction ref-lookup)
         (t/attach-instruction ref-new)
         (t/attach-instruction ref-peek)
+        (t/attach-instruction ref->vector)
 
         aspects/make-collectible
         aspects/make-equatable